pyproject.toml
[tool.poetry]
name = "superorga"
version = "0.1.0"
description = ""
authors = ["quentalt <quentin.altieri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
networkx = "^3.3"
matplotlib = "^3.9.1.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
